+++ examples/telescope/components/index.tsx
@@ -1,0 +1,5 @@
+export * from './types';
+export * from './wallet-connect';
+export * from './user-info';
+export * from './astronaut';
+export * from './features';