+++ packages/create-cosmos-app/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.2](https://github.com/cosmology-tech/create-cosmos-app/compare/create-cosmos-app@0.3.1...create-cosmos-app@0.3.2) (2022-09-08)
+
+**Note:** Version bump only for package create-cosmos-app
+
+
+
+
+
+## [0.3.1](https://github.com/cosmology-tech/create-cosmos-app/compare/create-cosmos-app@0.3.0...create-cosmos-app@0.3.1) (2022-09-02)
+
+**Note:** Version bump only for package create-cosmos-app
+
+
+
+
+
+# [0.3.0](https://github.com/cosmology-tech/create-cosmos-app/compare/create-cosmos-app@0.2.4...create-cosmos-app@0.3.0) (2022-09-02)
+
+**Note:** Version bump only for package create-cosmos-app
+
+
+
+
+
+## [0.2.4](https://github.com/cosmology-tech/create-cosmos-app/compare/create-cosmos-app@0.2.3...create-cosmos-app@0.2.4) (2022-09-02)
+
+**Note:** Version bump only for package create-cosmos-app
+
+
+
+
+
+## [0.2.3](https://github.com/cosmology-tech/create-cosmos-app/compare/create-cosmos-app@0.2.2...create-cosmos-app@0.2.3) (2022-09-02)
+
+**Note:** Version bump only for package create-cosmos-app
+
+
+
+
+
 ## [0.2.2](https://github.com/cosmology-tech/create-cosmos-app/compare/create-cosmos-app@0.2.1...create-cosmos-app@0.2.2) (2022-08-27)
 
 **Note:** Version bump only for package create-cosmos-app